--- server/server_metrics.csv
@@ -1,18 +1,0 @@
-9051172194
-4846729760
-2721770619
-1361728133
-888184956
-418860382
-239897707
-110647593
-72090521
-53119159
-44146603
-33701458
-37697330
-35770164
-50973571
-42398490
-51103578
-46866174
